Ajustados endpoints para melhor desempenho
